Move authorization-related elements to the 'auth' element group folder
