--- angular/components/elements/auth-console/auth-console.component.scss
@@ -1,9 +1,0 @@
-// auth-console.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-auth-console {
-    @include flex(center, center, column);
-}
