--- angular/components/elements/auth-popup/auth-popup.component.scss
@@ -1,17 +1,0 @@
-// auth-popup.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-auth-popup {
-    @include flex(center, center, row);
-    @extend %soft-border-radius;
-    @extend %softer-raise-above-box-shadow;
-    background: radial-gradient(at center, $pristine-white, $alabaster-stone);
-    padding: $one-percent;
-    margin: $one-percent 0;
-    ss-title {
-        text-decoration: underline;
-    }
-}
