+++ angular/components/elements/element-group-auth/auth-console/auth-console.component.scss
@@ -1,0 +1,9 @@
+// auth-console.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+auth-console {
+    @include flex(center, center, column);
+}
