+++ angular/components/elements/element-group-auth/auth-popup/auth-popup.component.ts
@@ -1,0 +1,18 @@
+// auth-popup.component.ts 
+import {Component, Input} from "@angular/core";
+import {TagType} from "../../../../models/html/TagType";
+import {PossibleAuthPopup} from "../../../../models/auth/types";
+
+@Component({
+    selector: 'auth-popup',
+    templateUrl: './auth-popup.component.html',
+    styleUrls: ['./auth-popup.component.css']
+})
+export class AuthPopupComponent {
+    @Input() protected authPopup: PossibleAuthPopup;
+    constructor() {
+        
+    }
+
+    protected readonly TagType = TagType;
+}
