+++ angular/components/elements/element-group-auth/auth-type-option/auth-type-option.component.html
