+++ angular/components/elements/element-group-auth/auth-type-slider/auth-type-slider.component.html
