+++ angular/components/elements/element-group-auth/auth-type-slider/auth-type-slider.component.scss
@@ -1,0 +1,13 @@
+// auth-type-slider.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+auth-type-slider {
+    @include flex(center, center, row);
+    padding: $half-percent $one-percent;
+    @extend %soft-border-radius;
+    background-color: $pristine-white;
+    margin: $one-percent 0;
+}
