+++ angular/components/elements/element-group-auth/login-inputs/login-inputs.component.scss
@@ -1,0 +1,10 @@
+// login-inputs.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+login-inputs {
+    @include flex(center, center, column);
+    width: 100%;
+}
