+++ angular/components/elements/element-group-auth/service-terms-input/service-terms-input.component.scss
@@ -1,0 +1,18 @@
+// service-terms-input.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+service-terms-input {
+    @include flex(center, center, row);
+    width: 100%;
+    margin: $quarter-percent 0;
+    ss-title {
+        text-decoration: underline;
+        text-decoration-color: $light-charcoal;
+        h5 {
+            text-shadow: none;
+        }
+    }
+}
