+++ angular/components/elements/element-group-auth/signup-console/signup-console.component.html
