--- angular/components/elements/signup-input/signup-input.component.scss
@@ -1,20 +1,0 @@
-// signup-input.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-signup-input {
-    @include flex(center, center, row);
-    width: 100%;
-    margin: $quarter-percent 0;
-    ss-title {
-        @include flex(flex-start, center, row);
-        flex: 1;
-        text-align: left;
-        width: 100%;
-    }
-    ss-input {
-        flex: 1;
-    }
-}
