--- angular/components/elements/signup-inputs/signup-inputs.component.scss
@@ -1,10 +1,0 @@
-// signup-inputs.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-signup-inputs {
-    @include flex(center, center, column);
-    width: 100%;
-}
